create a BlockStoreMeta interface
