+++ core/server/worker/src/main/java/alluxio/worker/block/BlockStore.java
@@ -288,7 +288,7 @@ interface BlockStore extends SessionCleanable {
   BlockStoreMeta getBlockStoreMeta();
 
   /**
-   * Similar as {@link BlockStoreMeta#getBlockStoreMeta} except that this includes more information
+   * Similar as {@link BlockStoreMeta.Factory#getBlockStoreMeta} except that this includes more information
    * about the block store (e.g. blockId list). This is an expensive operation.
    *
    * @return full store metadata
